Extract server auth helpers
